+++ src/Space/Space.hpp
@@ -12,7 +12,7 @@
 class Space {
 public:
 	// Give Entity possibility to move in Space or derived class of Space
-	virtual bool moveTo(int entityId, int entityType, int newX, int newY) = 0;
+	virtual bool moveTo(int entityId, int entityType, double newX, double newY) = 0;
 
 	// Give Entity possibility to remove/consume in Space or derived class of Space
 	virtual void remove(int entityId, int entityType) = 0;
